+++ Changelog.md
@@ -1,5 +1,11 @@
 # Changelog 
 
+## Release 0.1.4
+
+- add collection filter to zenkit object
+- add workspace property to zenkit object to discover workspaces easily
+- rename private workspaces field in zenkit to _workspaces
+
 ## Release 0.1.3
 
 - Set Zenkit-API-Key in axios requests only when requests go to Zenkit BASE_URL